pyproject.toml
[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nettools"
version = "1.0.0"
description = "Professional Network Analysis Tool"
authors = [{name = "DrDKuiper", email = "drdkuiper@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Networking :: Monitoring",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "PyQt6>=6.4.0",
    "psutil>=5.9.0",
    "matplotlib>=3.6.0",
    "requests>=2.28.0",
    "ping3>=4.0.0",
    "dnspython>=2.2.0",
    "netifaces>=0.11.0",
    "speedtest-cli>=2.1.3",
    "darkdetect>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "pytest>=7.0.0",
    "pytest-qt>=4.0.0",
]
build = [
    "pyinstaller>=5.7.0",
    "auto-py-to-exe>=2.20.0",
]

[project.urls]
Homepage = "https://github.com/DrDKuiper/NetTools"
"Bug Reports" = "https://github.com/DrDKuiper/NetTools/issues"
"Source Code" = "https://github.com/DrDKuiper/NetTools"

[project.scripts]
nettools = "main:main"
